learning/requirements.txt
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
yt-dlp==2024.1.1
openai-whisper==20231117
chromadb==0.4.22
minio==7.2.3
pillow==10.2.0
ffmpeg-python==0.2.0
numpy==1.26.3
dashscope==1.17.0
